Fix Duration Marshall methods to handle millisecond durations
Currently the Duration.MarshallText() and Duration.UnmarshallText()
methods format millisecond values as nanoseconds. This is because
they Duration type is managed in nanoseconds, but when the methods
deal with a millisecond, they don't multiply the value by
time.Millisecond. The corrosponding tests also have the same
behaviour.

This commit changes the tests and methods to essentially multiply
by time.Millisecond, but also modifies the MarshallText() method
to be slightly more concise with its handling of other time units.