Change wrapper attributes to bool pointer (#18)
Not all states can be captured with bool. Specifically followAdditionalWrappers has a default of true, but this cannot be detected with a boolean. For example, having the attribute not present and having the attribute with the value false will both result in the parsing returning a boolean with value false.

Using a bool pointer allows to dissociate false from missing value.